fix(utils): recognize negative numbers as integers in NumberTypeTester
The integer-detection regex ("[0-9]+") didn't allow a leading '-', so
negative numbers were misclassified as decimals. On JS, this made the
DSL termify e.g. -1 as a real (-1.0) instead of an integer, which was
masking as a JS-only quirk in an io-lib test; that test is restored to
its cross-platform form now that the underlying bug is fixed.
